Split trace/span id tests in different files
