+++ consumer/pdata/trace_test.go
@@ -95,26 +95,6 @@ func TestSpanCountWithEmpty(t *testing.T) {
 	}}.SpanCount())
 }
 
-func TestSpanID(t *testing.T) {
-	sid := InvalidSpanID()
-	assert.EqualValues(t, [8]byte{}, sid.Bytes())
-	assert.True(t, sid.IsEmpty())
-
-	sid = NewSpanID([8]byte{1, 2, 3, 4, 4, 3, 2, 1})
-	assert.EqualValues(t, [8]byte{1, 2, 3, 4, 4, 3, 2, 1}, sid.Bytes())
-	assert.False(t, sid.IsEmpty())
-}
-
-func TestTraceID(t *testing.T) {
-	tid := InvalidTraceID()
-	assert.EqualValues(t, [16]byte{}, tid.Bytes())
-	assert.True(t, tid.IsEmpty())
-
-	tid = NewTraceID([16]byte{1, 2, 3, 4, 5, 6, 7, 8, 8, 7, 6, 5, 4, 3, 2, 1})
-	assert.EqualValues(t, [16]byte{1, 2, 3, 4, 5, 6, 7, 8, 8, 7, 6, 5, 4, 3, 2, 1}, tid.Bytes())
-	assert.False(t, tid.IsEmpty())
-}
-
 func TestSpanStatusCode(t *testing.T) {
 	td := NewTraces()
 	rss := td.ResourceSpans()
